Estimated rows are now reported as floats in the format of '%.3f' in plan descriptions

+++ community/cypher/cypher-compiler-2.2/src/main/scala/org/neo4j/cypher/internal/compiler/v2_2/pipes/DirectedRelationshipByIdSeekPipe.scala
@@ -25,11 +25,10 @@ import org.neo4j.cypher.internal.compiler.v2_2.planDescription.InternalPlanDescr
 import org.neo4j.cypher.internal.compiler.v2_2.planDescription.{NoChildren, PlanDescriptionImpl}
 import org.neo4j.cypher.internal.compiler.v2_2.symbols._
 import org.neo4j.cypher.internal.helpers.CollectionSupport
-import org.neo4j.graphdb.Relationship
 
 
 case class DirectedRelationshipByIdSeekPipe(ident: String, relIdExpr: EntityByIdRhs, toNode: String, fromNode: String)
-                                           (val estimatedCardinality: Option[Long] = None)
+                                           (val estimatedCardinality: Option[Double] = None)
                                            (implicit pipeMonitor: PipeMonitor)
   extends Pipe
   with CollectionSupport
@@ -63,5 +62,5 @@ case class DirectedRelationshipByIdSeekPipe(ident: String, relIdExpr: EntityById
 
   override def localEffects = Effects.READS_ENTITIES
 
-  def withEstimatedCardinality(estimated: Long) = copy()(Some(estimated))
+  def withEstimatedCardinality(estimated: Double) = copy()(Some(estimated))
 }
